requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
groq
redis
requests
